+++ attmap/pathex_attmap.py
@@ -13,6 +13,10 @@ __all__ = ["PathExAttMap"]
 class PathExAttMap(OrdAttMap):
     """ Used in pepkit projects, with Mapping conversion and path expansion """
 
+    def __getattribute__(self, item, expand=True):
+        res = super(PathExAttMap, self).__getattribute__(item)
+        return _safely_expand(res) if expand else res
+
     def __getattr__(self, item, default=None, expand=True):
         """
         Get attribute, accessing stored key-value pairs as needed.
@@ -33,7 +37,7 @@ class PathExAttMap(OrdAttMap):
                 # __getitem__ syntax, not attribute-access syntax.
                 raise AttributeError(item)
         else:
-            return expandpath(v) if expand else v
+            return _safely_expand(v) if expand else v
 
     def __getitem__(self, item, expand=True):
         """
@@ -45,7 +49,7 @@ class PathExAttMap(OrdAttMap):
         :raise KeyError: if the requested key is unmapped.
         """
         v = super(PathExAttMap, self).__getitem__(item)
-        return self._finalize_value(v) if expand else v
+        return _safely_expand(v) if expand else v
 
     def items(self, expand=False):
         """
@@ -92,28 +96,10 @@ class PathExAttMap(OrdAttMap):
         """
         return self._simplify_keyvalue(self.items(expand), dict)
 
-    def _finalize_value(self, v):
-        """
-        Make any modifications to a retrieved value before returning it.
-
-        This hook accesses an instance's declaration of value mutations, or
-        transformations. That sequence may be empty (the base case), in which
-        case any value is simply always returned as-is.
-
-        If an instances does declare retrieval modifications, though, the
-        declaration should be an iterable of pairs, in which each element's
-        first component is a single-argument predicate function evaluated on
-        the given value, and the second component of each element is the
-        modification to apply if the predicate is satisfied.
-
-        At most one modification function will be called, and it would be the
-        first one for which the predicate was satisfied.
-
-        :param object v: a value to consider for modification
-        :return object: the finalized value
-        """
-        return expandpath(v) if isinstance(v, str) else v
-
     @property
     def _lower_type_bound(self):
         return PathExAttMap
+
+
+def _safely_expand(x):
+    return expandpath(x) if isinstance(x, str) else x